requirements.txt
numpy>=2.0
pyarrow>=17.0
scipy>=1.14
spiceypy>=6.0
pyglm>=2.7
natsort>=8.4
matplotlib>=3.9
cartopy>=0.23
pandas>=2.2
pywavefront>=1.3
arrow>=1.3